Jqwik Support: Grouped restarted containers
A PropertyLifecyclyeContext does not provide the means to access the parent instance.
This makes it hard to access the parent test instance. The nested container is started
and stopped. Different to Jupiter, Jqwik traverses the tree of test elements bottom-up
and first runs the properties within a group.
